# src/main/java/ru/vsu/cs/sheina/online_gallery_backend/dto/field/IntIdRequestDTO.java
package ru.vsu.cs.sheina.online_gallery_backend.dto.field;

import lombok.Data;
import lombok.RequiredArgsConstructor;

@Data
@RequiredArgsConstructor
public class IntIdRequestDTO {

    Integer id;
}